add methods

+++ carApp/carApp/carApp/src/main/java/com/dolittle/carApp/carMaintenance/CarServiceImpl.java
@@ -1,10 +1,10 @@
 package com.dolittle.carApp.carMaintenance;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
-import com.dolittle.carApp.carMaintenance.model.CarTO;
+import com.dolittle.carApp.employeeMaintenance.EmployeeEntity;
+import com.dolittle.carApp.model.CarTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,22 +20,46 @@ public class CarServiceImpl implements CarService {
     }
 
     @Override
-    public void createCar(CarEntity car) {
+    public void addCar(CarEntity car) {
         carDAO.save(car);
     }
     @Override
     public void updateCar(Long id, CarEntity car) {
         carDAO.save(car);
     }
+
     @Override
-    public void deleteCar(Long id) {
-        carDAO.deleteById(id);
+    public void assignEmployee(EmployeeEntity employeeId, CarTO car) {
+        car.getEmployees().add(employeeId);
     }
 
+    @Override
+    public void deleteCar(Long id) {
+        if(carDAO.existsById(id)) {
+            carDAO.deleteById(id);
+        }
+        else {
+            throw new IllegalArgumentException("Car doesnt exist!");
+        }
+    }
+
+    @Override
+    public List<CarTO> getCarsByModel(String model) {
+        return carDAO.findAll().stream()
+                .map(this::mapToCarTO).filter(c -> c.getModel().equals(model))
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public List<CarTO> getCarsByEmployee(Long employeeId) {
+        return carDAO.findAll().stream()
+                .map(this::mapToCarTO).filter(c -> c.getEmployees().contains(employeeId))
+                .collect(Collectors.toList());
+    }
 
     private CarTO mapToCarTO(CarEntity carEntity) {
         return new CarTO(carEntity.getId(), carEntity.getModel(), carEntity.getYearOfProduction(),
                 carEntity.getColor(), carEntity.getEngineCapacity(), carEntity.getEnginePower(),
-                carEntity.getMileage());
+                carEntity.getMileage(), carEntity.getEmployees());
     }
 }
